backend/orders/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy
kafka-python
pydantic>=2.0
grpcio
grpcio-tools
requests
python-jose[cryptography]